% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd_detailed.R
\name{all_na}
\alias{all_na}
\title{Check if vector is all NA}
\usage{
all_na(data)
}
\arguments{
\item{data}{vector of data.frame}
}
\value{
logical
}
\description{
Check if vector is all NA
}
\examples{
rep(NA,4) |> all_na()
}
