% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.R
\name{possibly_roman}
\alias{possibly_roman}
\title{Test if vector can be interpreted as roman numerals}
\usage{
possibly_roman(data)
}
\arguments{
\item{data}{character vector}
}
\value{
logical
}
\description{
Test if vector can be interpreted as roman numerals
}
\examples{
sample(1:100, 10) |>
  as.roman() |>
  possibly_roman()
sample(c(TRUE, FALSE), 10, TRUE) |> possibly_roman()
rep(NA, 10) |> possibly_roman()
}
