% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cast.R
\name{read_input}
\alias{read_input}
\title{Flexible file import based on extension}
\usage{
read_input(file, consider.na = c("NA", "\\"\\"", ""), sheet = 1, ...)
}
\arguments{
\item{file}{file name}

\item{consider.na}{character vector of strings to consider as NAs}

\item{sheet}{sheet number}

\item{...}{magic dots to allow overflow. Ignored for now.}
}
\value{
tibble
}
\description{
Flexible file import based on extension
}
\examples{
read_input("https://raw.githubusercontent.com/agdamsbo/cognitive.index.lookup/main/data/sample.csv")
}
