% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redcap.R
\name{read_redcap_labelled}
\alias{read_redcap_labelled}
\title{REDCap read function to preserve field labels and all factor levels}
\usage{
read_redcap_labelled(
  uri,
  token,
  records = NULL,
  fields = NULL,
  events = NULL,
  forms = NULL,
  split_forms = "all"
)
}
\arguments{
\item{uri}{REDCap database API uri}

\item{token}{API token}

\item{records}{records to download}

\item{fields}{fields to download}

\item{events}{events to download}

\item{forms}{forms to download}

\item{split_forms}{Whether to split "repeating" or "all" forms, default is
"all".}
}
\value{
data.frame or list
}
\description{
This works very much as `read_redcap_tables()` and might end up there
}
