% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd_detailed.R
\name{redcap_meta_default}
\alias{redcap_meta_default}
\title{Default column names of a REDCap data dictionary}
\usage{
redcap_meta_default(...)
}
\arguments{
\item{...}{ignored for now}
}
\value{
character vector
}
\description{
Default column names of a REDCap data dictionary
}
\examples{
dput(redcap_meta_default())
}
