% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2dd.R
\name{replace_curly_quote}
\alias{replace_curly_quote}
\title{Replace curly apostrophes and quotes from word}
\usage{
replace_curly_quote(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Copied from textclean, which has not been updated since 2018 and is not
on CRAN. Github:https://github.com/trinker/textclean
}
