% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sanitize_split}
\alias{sanitize_split}
\title{Sanitize list of data frames}
\usage{
sanitize_split(
  l,
  generic.names = c("redcap_event_name", "redcap_repeat_instrument",
    "redcap_repeat_instance"),
  drop.complete = TRUE,
  drop.empty = TRUE
)
}
\arguments{
\item{l}{A list of data frames.}

\item{generic.names}{A vector of generic names to be excluded.}

\item{drop.complete}{logical to remove generic REDCap variables indicating
instrument completion. Default is TRUE.}

\item{drop.empty}{logical to remove variables with only NAs Default is TRUE.}
}
\value{
A list of data frames with generic names excluded.
}
\description{
Removing empty rows
}
