% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.R
\name{set_attr}
\alias{set_attr}
\title{Set attributes for named attribute. Appends if attr is NULL}
\usage{
set_attr(data, label, attr = NULL, overwrite = FALSE)
}
\arguments{
\item{data}{vector}

\item{label}{label}

\item{attr}{attribute name}

\item{overwrite}{overwrite existing attributes. Default is FALSE.}
}
\value{
vector with attribute
}
\description{
Set attributes for named attribute. Appends if attr is NULL
}
