% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cast.R
\name{shiny_cast}
\alias{shiny_cast}
\title{Launch the included Shiny-app for database casting and upload}
\usage{
shiny_cast(...)
}
\arguments{
\item{...}{Arguments passed to shiny::runApp()}
}
\value{
shiny app
}
\description{
Wraps shiny::runApp()
}
\examples{
# shiny_cast()

}
